# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "description": "Login User Here",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/criar-evento": {
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Create event",
                "description": "responsável por criar um evento",
                "operationId": "event",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "status",
                                    "data_inicio",
                                    "data_prazo"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "data_inicio": {
                                        "type": "timestamp"
                                    },
                                    "data_prazo": {
                                        "type": "timestamp"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/eventos": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar eventos",
                "description": "Listar todos os eventos",
                "operationId": "eventall",
                "responses": {
                    "200": {
                        "description": "Eventos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/eventos/{id}": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar um evento específico",
                "description": "Listar um evento específico",
                "operationId": "eventabyid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Título do evento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "teste"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eventos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Evento não encontrado ou você não tem permissão para vê-lo",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/excluir-evento/{id}": {
            "delete": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Delete um evento",
                "operationId": "ApiV1DeleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Título do evento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "teste"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Evento não encontrado ou você não tem permissão para excluí-lo"
                    },
                    "200": {
                        "description": "Evento deletado com sucesso"
                    }
                }
            }
        },
        "/editar-evento/{nomeEvento}": {
            "put": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Put event",
                "description": "Atualizado o evento com base em seu id",
                "operationId": "eventput2",
                "parameters": [
                    {
                        "name": "nomeEvento",
                        "in": "path",
                        "description": "Nome do evento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "teste"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "data_inicio": {
                                        "type": "timestamp"
                                    },
                                    "data_prazo": {
                                        "type": "timestamp"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento editado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Evento não encontrado ou você não tem permissão para editá-lo"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}